Created RadiButton of Kassa

+++ src/BurgerMenu-Desktop/ViewModels/Products/ProductWithDetailsViewModel.cs
@@ -10,7 +10,7 @@ public class ProductWithDetailsViewModel
 {
     public long id {  get; set; }   
     public string product_name { get; set; } = string.Empty;
-    public long quantity { get; set; }
+    public int quantity { get; set; }
     public float StartingPrice { get; set; }
     public float SoldPrice{ get; set; }
     public long BarCode { get; set; }
